requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-extensions>=3.1,<3.3
pydotplus>=2.0.2
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.20.2,<0.21